Follower/Learner can auto reconnect, when leader is shutdown. (#1879)
### What problem does this PR solve?

1. Shutdown leader
2. The follower/learner should issue error message.
3. Restart leader
4. The follower and learner will auto reconnect to the leader
5. ADMIN SHOW NODE 'node_name';
6. Once follower / learners register to leader, leader will also try to
connect follower / learner servers.
7. leader send logs to follower / learners before registration finished.
8. TODO: leader send logs to follower and learner after TXN finished.

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)

---------

+++ src/network/http_server.cpp
@@ -3117,6 +3117,9 @@ public:
 namespace infinity {
 
 void HTTPServer::Start(const String& ip_address, u16 port) {
+    if(started_) {
+        return ;
+    }
 
     WebEnvironment::init();
 
@@ -3187,13 +3190,17 @@ void HTTPServer::Start(const String& ip_address, u16 port) {
 
     fmt::print("HTTP server listen on {}: {}\n", ip_address, port);
 
+    started_ = true;
+
     server_->run();
 }
 
 void HTTPServer::Shutdown() {
-
-    server_->stop();
-    WebEnvironment::destroy();
+    if(started_) {
+        server_->stop();
+        WebEnvironment::destroy();
+        started_ = false;
+    }
 }
 
 } // namespace infinity
